Connection metadata editable in properties view
This patch adds the ability to display and edit the metadata of a
connection in the properties view. To access the metadata, select
any trace link in the trace matrix view. The comment field can be
edited directly. Changes are not saved until the trace model is
changed since the automatic storing of the metadata model on changes
is not yet implemented.

Change-Id: I6da8c529ff40eb2b0ac9d78baf35bb394b660b26