PersistenceAdapter no longer prints stacktrace
The persistence adapter used to print a stacktrace to the console
whenever a model could not be loaded. This was not useful during
test runs since this situation is expected there (the models are
held in memory, not on disk). This patch therefore downgrades the
error to a warning and only prints out a short message.

Change-Id: Ie08e20c9e975fc381df7956845d7fa90db687423