[Fix] Matrix view now retrieves connection properly
This patch fixes an issue where the TraceabilityDataProvider
did not use the row to retrieve the Connection of a cell selected
by the user. While this error has not manifested as all matrices
are symmetrical at this point, this could potentially cause issues
if the matrix is populated differently in the future.

Change-Id: I76c55a8b5ea762dc4eaf67191a471f4d34d2044b