Added namespace to the artifact and trace metamodel
The commit adds explicit namespaces containing a version number to
the xcore metamodels of the artifact model and the trace model. This
allows us to detect which version of the metamodels any model we load
is based on and transform them to the current version if necessary.
Whenever a model is loaded from disk, we can perform the version check
and then transform if necessary before loading the models with the
current metamodel version.

Change-Id: I94d06de3ef545484b964e2792571abfd18792cc2