Matrix shows origins in rows, targets in columns
The matrix view previously always showed a symmetrical view of the
artifacts, meaning that rows and columns were identical. In addition,
they were sorted by their appearance in the trace links that were
shown. This led to a hodgepodge of artifacts and made it harder to
identify the relevant parts.

This patch changes the behaviour as follows:
* All origins related to artifacts selected somewhere in Eclipse show
  up as rows.
* All targets from the trace links that have the selected elements as
  origins show up in columns.
* The artifacts in both columns and rows are sorted by the class name
  of their respective artifact handlers. This effectively groups
  artifacts of the same type together.

This makes it easier to see if, e.g., all selected requirements are
covered by test cases or if all failures are associated to hazards.

Change-Id: I6274290127af44ad6c9c1932481d1e57d0bbb6b4