Office view now using dependencies from Orbit
The office view shipped with the POI libraries. This patch changes
this and the office view is now using the dependencies from the
target platform which gets POI from Orbit.

One caveat: the org.apache.xmlbeans library contains classes
in the org.w3c.dom and the javax.xml.parsers packages. Since
these are also provided by the JDK and Java 9+ does not allow
libraries shadowing JDK-provided classes on the classpath, I
needed to set the source compliance level to Java 1.8 in the
MANIFEST.MF to make this work. Right now, this is not an issue,
but if we want to use any Java 9+ features in the future, we need
to revisit this.

Change-Id: I57dd6f5c66897806e2f81f5bbc37b0ee42cd1eec