Add a trace creation listener
This patch adds a listener that is triggered whenever a new trace
is created. Listeners can be registered via an extension point.
All registered listeners are called after the trace has been added
to the trace model.

As an example for this, a listener is added that records some
generic metadata, namely the current data and the user name of the
person creating the link.

Change-Id: I9595f4cafc089936384eff5e35dc361cec71a34e