[Fix] Metadata now removed when trace is deleted
The introduction of metadata makes it necessary to do some
housekeeping when deleting a trace link. Since the metadata
model references the trace links, the metadata for a link
also has to be deleted when the trace link itself is deleted.
Failing to do so leads to a
org.eclipse.emf.ecore.xmi.DanglingHREFException when storing
the models.

This patch adds a removeMetadata() method to IMetadataAdapter.
It also calls this new method in ITraceabilityInformationAdapter.
deleteTrace(). This ensures that no dangling references to links
that have been deleted from the trace model remain.

Change-Id: I6b449cdfb6dd6c87ef7fd9adb654b852401c46f8