Metadata now stored on update in properties source
Metadata can be shown in a trace links properties view by selecting
the trace link in the matrix viewer. It was already possible to
change the values there, but changes were not yet persisted. This
happens now by storing the models in the current resource set when
one of the properties is set.

Change-Id: Ia3237110f6bfaf5beaf67a4baf1b3bbb89fa5642