Selections in matrix view now more robust
The matrix view distinguishes selections of columns, rows, and cells.
When a matrix only has one row or one column, the previous
implementation did not allow selecting cells due to the way the
selection was determined. This patch changes this and now robustly
allows the selection of the entire row, the entire column, or
individual cells.

Change-Id: If6dfaf29f3b16d1ed5e5ffc79723dcceaf039089