Excel files now shown in table in OfficeView
Adding NatTable for OfficeView so that the user
can see the contents of an Excel file in a table
in the Office Viewer, and get an idea of the
contents of the entire file. The NatTable is populated
by performing drag and drop action, and will allow
the user to create trace links for the data displayed.

Signed-off-by: Mihaela Grubii <laborator2test@gmail.com>
Change-Id: I675558b5d5391df40fffe48d58354e30811902f4