Trace models now use UUIDs to reference elements
The trace  model referred to artifacts using the standard EMF
approach of the position in the tree. Likewise, metadata entries
referred to the traces using the same mechanism. With this commit,
this is switched to referencing by UUID. This makes the models more
robust to changes that go not through EMF (e.g., manual editing).

Change-Id: If50db274c92486160b6f8b43adb702190677cc94