Removed deprecated target definitions
This commit removes target definitions for Eclipse 4.10 to 4.16. At the
same time, it sets the default target definition to 4.17. The reason for
this change is that it has become more and more complicated to maintain
old versions which are still based on Java 8 and update sites with
features compatible with these old versions are slowly disappearing.

Change-Id: I3cb788e6bda11a9c0b50f81bcfa15e3d04d06a89