[Releng] Added dependency to C/C++ tool chain to test suite
Since updating to a newer Eclipse target platform, we had issues with
the tests that involve CDT projects not completing. It turns out since
this is due to the fact that we did not have a dependency to a tool
chain which is required for the managed build projects that we are
using. This was not an issue within Eclipse, but Maven builds failed
because of this. This change adds a dependency to the GNU tool chain
which resolves the issue.