Add methods to compare EObject based on URI and name
So far, EMFHelper's functions have all been based on the name attribute.
For instance, when checking if two EObjects have a trace between them,
EMFHelper.hasSameIdentifier() was used to see if the EObjects in the
trace's origin and target were the ones we are looking for. This
behaviour breaks down when two elements have the same name, however.
Therefore, new methods that use the URI as the main way of comparing
objects have now been introduced.
Change-Id: I7091fef7567c390d364942dc942301a2efed8c87