Add support for reversing link direction
Eclipse Capra uses directed links, i.e., traceability links always need
to have an origin and a target. However, in many situations it is
helpful to traverse the trace model in the other direction, i.e. from
the targets to the origins, e.g., to perform rationale analysis.
Therefore, this commit introduces support to reverse the link directions
for isThereATraceBetween() and getConnectedElements(). If the new,
optional flag is set to true, origin and targets will be switched.
Change-Id: I7e5cc29fe4ef9704c668dadb2b5f63694926516f