Made arrow label for PlantUML view configurable
This commit adds a preference to show the full information of the
traceability link (including source and targets) on the arrows in
the PlantUML view or to only show the trace link type for a more
compact view. At the same time, it fixes a bug where the preference
store was not correctly loaded.

Change-Id: I24769b9164ddc727b2a630cfbe9015b8a3dcabb2