Do not update PlantUML view w/o selected element
Under certain circumstances, the PlantUML view switched to the
message to select an element to show the graph when the focus
changed to a different view. This was a problem since this also
happened when switching to the PlantUML view itself.

This commit changes this behaviour by not refreshing when no
element is selected. However, this is only a hacky fix since our
previous way of addressing this (via the supportsView() and
similar methods) should be the preferred way, but does not seem
to be currently triggered.

Change-Id: Ie41181bf81955c7bce01ec8dfa2edd463b5a4cb5