Fix deleting traces in matrix view not working
Under certain circumstances was it not possible to delete a trace
link from the matrix view. This is due to the fact that the action
relied on the global selection listener (which also picks up
selections in the matrix view) rather than the matrix views
selection layer. This patch fixes this behaviour.

Change-Id: If49745ed63ac4da1bc6399e3b59c8a5ca9ff3136