Use extension '.capra' for all models
Eclipse Capra used '.xmi' as the extension for the trace, artifact,
and metadata models. At the same time, it registered its own resource
factory for all file extensions. That meant that all file extensions
without their own registered resource factory were loaded into a
CapraResource. That in turn meant that UUIDs were enforced even for
models that do not use them, causing the references in the trace model
to be incorrect.

This patch changes this behaviour: all models used by Eclipse Capra
now use the '.capra' extension and the CapraResourceFactory is only
registered for this extensions. More details are available at:
https://www.eclipse.org/forums/index.php/m/1852682

Change-Id: Id79f427239069b35bfe5b0822d41b8d704da0c57