Fix: consider internal resolver when locating artifacts
After introducing the new structure for the artifact model, an internal
resolver can be used to identify an element within a file. However, when
looking for existing elements, the internal resolver was ignored. This
meant that handlers making use of the internal resolver did not have the
ability to create several artifact wrappers within the same file.

This commit fixes this problem by extending the
IArtifactMetaModelAdapter interface to include the internal resolver in
the getArtifact() method. It also adds a null-aware implementation in
the GenericArtifactMetaModelAdapter that makes sure that the existing
handlers continue to behave as expected.
Change-Id: Icb9a768c22c030fd02b696bd1d4b4ba0072c9814