Add handler for AsciiDoc files
This commit adds a handler for AsciiDoc files which uses the
Eclipse AsciiDoctor Editor available at:
https://github.com/de-jcup/eclipse-asciidoctor-editor/

At the moment, the handler allows tracing to all elements in an
AsciiDoc file that the AsciiDoctor Editor identifies as an "item".
These include headings as well as all manually input IDs.

AsciiDoctor Editor does not currently expose its classes as API.
This means that the handler is using non-API classes which might
cause problems at runtime. The workaround is to use a version of
AsciiDoctor Editor that exposes the necessary packages as
described in:
https://github.com/de-jcup/eclipse-asciidoctor-editor/issues/429

Change-Id: I66ec0895fe267548b6d6cbdc44471db303fb9007