Add new query-based API to retrieve trace links
The current API in the ITraceabilityInformationModelAdapter was getting
cumbersome since we kept overloading methods to add new parameters for
the retrieval of trace links from the trace model. This commit adds a
query-based API where a ConnectionQuery contains all relevant
parameters. The ITraceabilityInformationModelAdapter then needs to make
sure to deal with all of these parameters. The effect, however, is that
it is no longer necessary to have a number of overloaded methods in the
API.
Change-Id: Ibd9e12bd2c6439b981c49a61f2af8b05aadca9d4