Enable reverse direction for internal links in TIM adapter
While directly connected links and transitive links were already
retrieved honouring the reverseDirection flag of the connection query,
internal links were so far retrieved using the previous methods that
were not aware of directional reversal. This commit refactors the TIM
adapter so that the flag in the query is now honored for internal links
as well, at least as long as the handlers support this.
Change-Id: I9ac2044cd02dcf3cb304847e62c16063aa733b2e