Fix: Metadata deleted in same transaction as trace
Previously, a trace links metadata was deleted before the transaction
to delete the trace itself was executed. If the transaction had to
be rolled back for some reason, the metadata was therefore no longer
available, potentially leading to inconsistencies. With this patch,
the trace metadata is now deleted in the same transaction as the
trace itself, thus fixing this issue.

Change-Id: Id8c34fe65685d37f74efb6b2a75d7160f3a9a56c