Do not ask for trace link type if only one applies
Previously, Eclipse Capra always opened the dialog to choose a trace
link type for a new link, even if just one type was applicable. With
this change, the trace link type selection dialog will no longer
appear if only one type of trace link applies to the artifacts
that are currently in the selection view.

Change-Id: Ie90f9ee4cb83d66344edfc379108e7f48abef960