Added ability to check if artifact exists
The IArtifactHandler now contains a new method `doesArtifactExist()`
with a default implementation that checks if the resolved wrapper is not
null. Some of the handlers also have more sophisticated implementations:
the FileHandler, e.g., checks if the file exists and the CDTHandler and
JavaHandler both check if the referenced C/C++ or Java element is
actually in the source code.
Change-Id: I8c6eb8bcff8ce648fe45c2dd5e83041fffd80f93