Improved property sources for connections and metadata
The property sources for connections and metadata can be used to access
information about a trace link. This commit makes the different
properties accessible via a String as well as via the DescriptorIDs.
This means that other views can access the information, even if they
only have a String representation of the desired data field.

In addition, this commit makes the "Creation User" metadata information
accessible which was previously not exposed.
Change-Id: If82242433b275d6ecab25477f265afe44325604b