Made checking artifact existence selectable in matrix view
The matrix view now has a new toggle item in the menu that lets the user
choose whether the existence of linked artifacts should be checked or
not. This addresses a performance issue since checking existence of an
artefact can be a long-running process that should not be triggered too
often.
Change-Id: I7e15a3d4e5f67dd25f4f39eb21b85ddc790b12f0