Added check of artifact status: normal, renamed, removed
On top of the current possibility to check if an artifact exists (or if
it has been removed), it is now also possible to determine if an
artifact has been renamed. A default implementation that falls back to
an artifact existence check is implemented, but handlers can override it
to implement custom behaviour. This is demonstrated using the AsciiDoc
handler which can now also detect if an artifact is still at the same
location but has a different name.
Change-Id: Ia05f16338b6addaa1e80513dc3eadcbb12f6f93e