Fixed the delete quick fix
Some of the architectural changes of recent times never made it to the
quick fixes. In particular, the DeleteQuickFix class had some issues
that were related to setting the artifact URIs in the marker
incorrectly, only looking for traces in which the deleted artifact was
an origin, and not using transactions. This commit fixes these issues
and adds test cases for it.

Change-Id: I7a47d770fa068e81f0211fa5795e111d9f58c2b7