User now is printed in nicer format in generic TIM
The metadata for a trace link contains a user who created the link
which is automatically set by Eclipse Capra. Until now, the toString()
method used for this printed a not so nice representation of the user.
With this commit, it is now printed as "user <user@email>" which
corresponds to the standard format used for email addressess.

Change-Id: Id6aea2f890c2279926658cde3a01014b21d68e13