Fixed changing ID for RelatedTo links
The way the ID field for RelatedTo links was populated meant
that each time the field was read with the generated getter(),
the ID was set to a new value. The new implementation uses a
custom factory to avoid this and sets the ID when the trace link
is created.

Change-Id: I8168abada12383935ea641495a82e01108659f67