Added ability to execute quick fixes from trace matrix view
Any artifact that has an associated Eclipse Capra problem marker (e.g.,
because it has been renamed or deleted) now has a "Quick Fix" context
menu in the trace matrix view. From this menu, the relevant quick fixes
can be accessed and executed directly. The column or row for the
artifact needs to be selected for this to work at this point.
Change-Id: Ic72d251de1b207d9a603159a865582960f78e68d