AsciiDoc artifacts can now be identified by ID
The AsciiDoc handler now distinguishes two cases: if an artifact is of
type "inline anchor", it will use the anchor's ID to locate the artefact
in the AsciiDoc document. Otherwise, it will default back to the offset,
e.g., for headings etc. This is more robust since the item can be
located by ID even when it has moved in the document.
Change-Id: I75f57ac3b36d6d6ccb323cabf6114edac386986a